requirements.txt
Flask==2.2.2
fit_tool==0.9.13
requests==2.28.1
oauthlib==3.2.2
pyOpenSSL==22.1.0
Flask-Login==0.6.2
Flask-WTF==1.0.1
Flask-Bootstrap4==4.0.2